HUDクラスの削除と初期シーンの変更
- HUDクラス（HUD.cpp, HUD.h）をプロジェクトから削除
  - HUDに関連するすべてのメンバ関数、変数、依存関係を整理
- ビルド設定の警告レベルをLevel4からLevel2に変更
- MagEngine.vcxprojに新しいインクルードディレクトリを追加
- 初期シーンをDebugSceneからGamePlaySceneに変更
- Enemy.cppでM_PIを明示的にfloat型にキャスト
- ImGui設定ファイル(imgui.ini)のDockNode選択状態を更新

+++ application/HUD.cpp
@@ -386,7 +386,7 @@ void HUD::DrawRollScale(float rollAngle) {
 
 	// スケール目盛り（-60度から+60度まで30度間隔）
 	for (int angle = -60; angle <= 60; angle += 30) {
-		float radians = DegreesToRadians(angle);
+		float radians = DegreesToRadians(static_cast<float>(angle));
 		float tickLength = (angle == 0) ? 1.5f : 1.0f;
 
 		float outerX = sinf(radians) * radius;
